requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r requirements/requirements.in
celery-progress==0.1.2
    # via -r requirements/requirements.in
certifi==2022.5.18.1
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via -r requirements/requirements.in
coreschema==0.0.4
    # via coreapi
cryptography==37.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-stripe==2.6.1
    # via -r requirements/requirements.in
django==3.2.13
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-hijack
    #   djangorestframework
    #   jsonfield
django-allauth==0.50.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.6
    # via -r requirements/requirements.in
django-hijack==3.2.1
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via -r requirements/requirements.in
djangorestframework-api-key==2.2.0
    # via -r requirements/requirements.in
# idna==3.3
google-auth==2.8.0
google-auth-oauthlib==0.5.2
google-api-python-client==2.52.0
idna==2.9.0
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonfield==3.1.0
    # via dj-stripe
kombu==5.2.4
    # via celery
markupsafe==2.1.1
    # via jinja2
numpy==1.23.0
    # via pandas
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via redis
pandas==1.4.3
    # via -r requirements/requirements.in
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via pandas
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   celery
    #   django
    #   djangorestframework
    #   pandas
redis==4.3.1
    # via celery
requests==2.25.1
    # via
    #   coreapi
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.5.12
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.2
    # via django
stripe==3.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated
